Проведена реструкторизация классов (*Command) по вложенности Добавлено сохранение номера телефона клиента *Добавлен пункт в меню Отредактировано меню в классе BackToPetMenuCommand Добавлена константа GET_SHELTER_CLIENT_NUMBER Добавлены поля статуса, и сохранения номера в клиенте Добавлен enum ShelterClientMode "Этап 2" +
