+++ src/main/java/pro/sky/teamoneproject/commands/buttonsforpets/buttomsforpet/ListOfDocumentsCommand.java
@@ -1,4 +1,4 @@
-package pro.sky.teamoneproject.commands.buttonsforpets;
+package pro.sky.teamoneproject.commands.buttonsforpets.buttomsforpet;
 
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.model.Update;
@@ -11,19 +11,18 @@ import pro.sky.teamoneproject.commands.Command;
 import static pro.sky.teamoneproject.constant.ConstantsForShelter.*;
 
 @Component
-public class HomeImprovementTipsCommand extends Command {
+public class ListOfDocumentsCommand extends Command {
     @Autowired
     private TelegramBot telegramBot;
 
-    public HomeImprovementTipsCommand() {
-        super(HOME_IMPROVEMENT_TIPS);
+    public ListOfDocumentsCommand() {
+        super(LIST_OF_DOCUMENTS);
     }
 
     @Override
     public void action(Update update) {
         long chatId = update.message().chat().id();
-        String messageText = update.message().text();
-        SendMessage sendMessage = new SendMessage(chatId, "Выберите интересующий Вас раздел");
+        SendMessage sendMessage = new SendMessage(chatId, "Список документов:\n1...\n2...\n3...");
         sendMessage.replyMarkup(getReplyKeyboard());
         telegramBot.execute(sendMessage);
     }
@@ -34,9 +33,6 @@ public class HomeImprovementTipsCommand extends Command {
      */
     private ReplyKeyboardMarkup getReplyKeyboard() {
         String[][] keyboard = new String[][] {
-                {HOME_IMPROVEMENT_TIPS_YOUNG_PET},
-                {HOME_IMPROVEMENT_TIPS_OLD_PET},
-                {HOME_IMPROVEMENT_TIPS_INVALID_PET},
                 {BACK_TO_PET_MENU}
         };
