+++ src/main/java/pro/sky/teamoneproject/commands/returnbuttons/BackToPetMenuCommand.java
@@ -22,7 +22,6 @@ public class BackToPetMenuCommand extends Command {
     @Override
     public void action(Update update) {
         long chatId = update.message().chat().id();
-        String messageText = update.message().text();
         SendMessage sendMessage = new SendMessage(chatId, "Пожалуйста, ознакомьтесь со всеми пунктами!");
         sendMessage.replyMarkup(getReplyKeyboard());
         telegramBot.execute(sendMessage);
@@ -42,7 +41,7 @@ public class BackToPetMenuCommand extends Command {
                 {DOG_HANDLERS_RECOMMENDATIONS},
                 {RULES_OF_SERVICE},
                 {CALL_VOLUNTEER},
-                {BACK_TO_SHELTER_MENU}
+                {BACK_TO_SELECT_PETS}
         };
 
         return new ReplyKeyboardMarkup(keyboard, true, false, false);
