+++ src/main/java/pro/sky/teamoneproject/constant/ShelterClientMode.java
@@ -1,0 +1,7 @@
+package pro.sky.teamoneproject.constant;
+
+public enum ShelterClientMode {
+    DEFAULT,
+    GET_PHONE_NUMBER,
+    SEND_PET_INFO
+}
