+++ src/main/java/pro/sky/teamoneproject/entity/ShelterClient.java
@@ -1,32 +1,37 @@
 package pro.sky.teamoneproject.entity;
 
 import jakarta.persistence.*;
+import pro.sky.teamoneproject.constant.ShelterClientMode;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
+
+import static pro.sky.teamoneproject.constant.ShelterClientMode.DEFAULT;
 
 @Entity
 public class ShelterClient {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String username;
     private Long chatId;
     @ManyToOne
     private Shelter selectedShelter;
+    private ShelterClientMode selectedMode;
     private LocalDateTime lastTimeAppeal;
-
     @OneToMany(fetch = FetchType.LAZY)
     public List<Pet> pets;
     private int petViewPageNumber;
-
+    private String phoneNumber;
     public ShelterClient() {}
 
     public ShelterClient(Long id, String username, Long chatId) {
         this.id = id;
         this.username = username;
         this.chatId = chatId;
+        this.selectedMode = DEFAULT;
     }
 
     public Long getId() {
@@ -61,6 +66,11 @@ public class ShelterClient {
         this.selectedShelter = selectedShelter;
     }
 
+    public ShelterClientMode getSelectedMode() {
+        if (selectedMode==null) selectedMode=DEFAULT;
+        return selectedMode;
+    }
+    public void setSelectedMode(ShelterClientMode selectedMode) { this.selectedMode = selectedMode; }
     public LocalDateTime getLastTimeAppeal() {
         return lastTimeAppeal;
     }
@@ -77,10 +87,20 @@ public class ShelterClient {
         this.petViewPageNumber = petViewPageNumber;
     }
 
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof ShelterClient that)) return false;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        ShelterClient that = (ShelterClient) o;
 
         if (petViewPageNumber != that.petViewPageNumber) return false;
         if (!Objects.equals(id, that.id)) return false;
@@ -88,9 +108,11 @@ public class ShelterClient {
         if (!Objects.equals(chatId, that.chatId)) return false;
         if (!Objects.equals(selectedShelter, that.selectedShelter))
             return false;
+        if (selectedMode != that.selectedMode) return false;
         if (!Objects.equals(lastTimeAppeal, that.lastTimeAppeal))
             return false;
-        return Objects.equals(pets, that.pets);
+        if (!Objects.equals(pets, that.pets)) return false;
+        return Objects.equals(phoneNumber, that.phoneNumber);
     }
 
     @Override
@@ -99,9 +121,11 @@ public class ShelterClient {
         result = 31 * result + (username != null ? username.hashCode() : 0);
         result = 31 * result + (chatId != null ? chatId.hashCode() : 0);
         result = 31 * result + (selectedShelter != null ? selectedShelter.hashCode() : 0);
+        result = 31 * result + (selectedMode != null ? selectedMode.hashCode() : 0);
         result = 31 * result + (lastTimeAppeal != null ? lastTimeAppeal.hashCode() : 0);
         result = 31 * result + (pets != null ? pets.hashCode() : 0);
         result = 31 * result + petViewPageNumber;
+        result = 31 * result + (phoneNumber != null ? phoneNumber.hashCode() : 0);
         return result;
     }
 }
