--- src/main/java/pro/sky/teamoneproject/service/RulesForPetService.java
@@ -1,10 +1,0 @@
-package pro.sky.teamoneproject.service;
-
-import pro.sky.teamoneproject.entity.RuleForPet;
-
-public interface RulesForPetService {
-
-    RuleForPet addRule(String text);
-
-    RuleForPet getRule(Long id);
-}
